+++ packages/client-common/src/connection.ts
@@ -11,7 +11,7 @@ export interface ConnectionParams {
   password: string
   database: string
   clickhouse_settings: ClickHouseSettings
-  logWriter: LogWriter
+  log_writer: LogWriter
   application_id?: string
   additional_headers?: Record<string, string>
 }