+++ packages/client-node/__tests__/unit/node_client.test.ts
@@ -3,20 +3,20 @@ import { createClient } from '../../src'
 
 describe('[Node.js] createClient', () => {
   it('throws on incorrect "host" config value', () => {
-    expect(() => createClient({ host: 'foo' })).toThrowError(
-      'Configuration parameter "host" contains malformed url.'
+    expect(() => createClient({ url: 'foo' })).toThrowError(
+      'Client URL is malformed.'
     )
   })
 
   it('should not mutate provided configuration', async () => {
-    const config: BaseClickHouseClientConfigOptions<unknown> = {
-      host: 'http://localhost',
+    const config: BaseClickHouseClientConfigOptions = {
+      url: 'http://localhost',
     }
     createClient(config)
     // initial configuration is not overridden by the defaults we assign
     // when we transform the specified config object to the connection params
     expect(config).toEqual({
-      host: 'http://localhost',
+      url: 'http://localhost',
     })
   })
 })